Feito os testes da entidade Cart e CartItem, onde foi validado os endpoint referente a inserção de produtos e deleção dos mesmos, e sanado o problema de Estoque e Reserva do Product, onde agora, ao realizar o a deleção dos items de um carrinho, automaticamente retorna os valores para o Product

+++ src/main/java/com/example/ecommerce/modules/cart/service/CartService.java
@@ -5,9 +5,12 @@ import com.example.ecommerce.modules.cart.exception.CartException;
 import com.example.ecommerce.modules.cart.mapper.CartMapper;
 import com.example.ecommerce.modules.cart.model.Cart;
 import com.example.ecommerce.modules.cart.model.CartEnum;
+import com.example.ecommerce.modules.cart.model.CartItem;
 import com.example.ecommerce.modules.cart.repository.CartRepository;
 import com.example.ecommerce.modules.customers.model.Customers;
 import com.example.ecommerce.modules.customers.repository.CustomerRepository;
+import com.example.ecommerce.modules.product.model.Product;
+import com.example.ecommerce.modules.product.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
@@ -17,9 +20,11 @@ import java.util.List;
 public class CartService extends CartMapper {
     private final CartRepository cartRepository;
     private final CustomerRepository customerRepository;
-    public CartService(CartRepository cartRepository,  CustomerRepository customerRepository) {
+    private final ProductRepository productRepository;
+    public CartService(CartRepository cartRepository,  CustomerRepository customerRepository,  ProductRepository productRepository) {
         this.cartRepository = cartRepository;
         this.customerRepository = customerRepository;
+        this.productRepository = productRepository;
     }
 
     public Cart createCart(Integer customerId) {
@@ -64,15 +69,38 @@ public class CartService extends CartMapper {
     public Cart clearCart(Integer cartId) {
         Cart cart = cartRepository.findById(cartId).orElseThrow(() ->
                 new CartException("Carrinho não encontrado"));
+        for (CartItem item : cart.getItems()) {
+            Product product = item.getProduct();
+
+            product.setQuantidadeEstoque(
+                    product.getQuantidadeEstoque() + item.getQuantidade()
+            );
+
+            product.setQuantidadeReservada(
+                    product.getQuantidadeReservada() - item.getQuantidade()
+            );
+
+            productRepository.save(product);
+        }
+
         cart.getItems().clear();
         cart.setValorTotal(0f);
+
         cartRepository.save(cart);
+
         return cart;
     }
+
     public void deleteCart(Integer cartId) {
         Cart cart = cartRepository.findById(cartId).orElseThrow(
                 () -> new CartException("Carrinho não existente")
         );
+        if (!cart.isStatus().equals(CartEnum.ATIVO)) {
+            throw new CartException("Só é possível deletar carrrinhos com Status Ativo!");
+        }
+        if (!cart.getItems().isEmpty()) {
+            throw new CartException("Só é possível deletar carrinho se não houver nenhum Item dentro dele.");
+        }
         cartRepository.delete(cart);
     }
 }
